Move seeder org repository extensions out of Scenes namespace
Relocate OrganizationRepositoryExtensions to a dedicated Extensions folder and mark it internal, addressing PR review feedback that a helper class did not belong in the Scenes namespace and need not be public.
